feature: move assign task to miranum platform
