--- connect/examples/process-integration-example/src/main/java/io/miragon/miranum/platform/adapter/in/task/dto/AssignTaskDto.java
@@ -1,13 +1,0 @@
-package io.miragon.miranum.platform.adapter.in.task.dto;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class AssignTaskDto {
-    private String assignee;
-}
