+++ connect/examples/process-integration-example/src/main/java/io/miragon/miranum/platform/application/port/in/WorkOnUserTaskInPort.java
@@ -8,8 +8,4 @@ public interface WorkOnUserTaskInPort {
 
     void completeUserTask(String user, String taskId, Map<String, Object> payload) throws TaskAccessDeniedException;
 
-    void assignUserTask(String user, String taskId, String assignee) throws TaskAccessDeniedException;
-
-    void unassignUserTask(String user, String taskId) throws TaskAccessDeniedException;
-
 }
