+++ platform/engine/task/src/main/java/io/miragon/miranum/platform/tasklist/adapter/in/task/TaskController.java
@@ -1,7 +1,9 @@
 package io.miragon.miranum.platform.tasklist.adapter.in.task;
 
 import io.miragon.miranum.platform.security.authentication.UserAuthenticationProvider;
+import io.miragon.miranum.platform.tasklist.adapter.in.task.dto.AssignTaskDto;
 import io.miragon.miranum.platform.tasklist.application.port.in.UserTaskQuery;
+import io.miragon.miranum.platform.tasklist.application.port.in.WorkOnTaskUseCase;
 import io.miragon.miranum.platform.tasklist.domain.Task;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
@@ -10,10 +12,7 @@ import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
@@ -25,6 +24,7 @@ import java.util.List;
 public class TaskController {
 
     private final UserTaskQuery userTaskQuery;
+    private final WorkOnTaskUseCase workOnTaskUseCase;
     private final UserAuthenticationProvider authenticationProvider;
 
     @GetMapping("/user")
@@ -44,6 +44,21 @@ public class TaskController {
             @Parameter(name = "taskId", description = "A task id string used during search with the task string.", in = ParameterIn.PATH) @Valid @PathVariable(value = "taskId", required = false) String taskId
     ) {
         return userTaskQuery.getTask(taskId, authenticationProvider.getLoggedInUser());
+    }
+
+    @PostMapping("/{taskId}/assign")
+    public void assignTask(
+            @Parameter(name = "taskId", description = "A task id string used during search with the task string.", in = ParameterIn.PATH) @Valid @PathVariable(value = "taskId", required = false) String taskId,
+            @Valid @RequestBody AssignTaskDto assignTaskDto
+    ) {
+        this.workOnTaskUseCase.assignUserTask(authenticationProvider.getLoggedInUser(), taskId, assignTaskDto.getAssignee());
+    }
+
+    @PostMapping("/{taskId}/unassign")
+    public void unassignTask(
+            @Parameter(name = "taskId", description = "A task id string used during search with the task string.", in = ParameterIn.PATH) @Valid @PathVariable(value = "taskId", required = false) String taskId
+    ) {
+        this.workOnTaskUseCase.unassignUserTask(authenticationProvider.getLoggedInUser(), taskId);
     }
 
 }
